Refactor: Align wordBreak folder structure and distribute code
I moved the wordBreak problem files (`wordBreak.ts`, `wordBreak.test.ts`) into a dedicated directory `packages/backend/src/problem/free/wordBreak/`.

I created standard placeholder files (`code/index.ts`, `code/solution.ts`, `groups.ts`, `problem.ts`, `steps.ts`, `testcase.ts`, `types.ts`, `variables.ts`) within the new directory.

I distributed the code from the original `wordBreak.ts` into the appropriate new files:
- `wordBreak` function logic moved to `code/index.ts`
- `Problem` definition moved to `problem.ts`
- `code` string moved to `code/solution.ts`
- Type definitions (`WordBreakInput`) moved to `types.ts`
- `title` and `getInput` moved to `problem.ts`

I renamed `wordBreak.test.ts` to `index.test.ts`.

I updated relative import paths within all affected files (`index.test.ts`, `problem.ts`, `code/index.ts`) to reflect the new structure.

I deleted the original, now empty, `wordBreak.ts` file.

This aligns the wordBreak problem with the standard project structure.
